backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
requests==2.32.3
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
mediapipe==0.10.14 
plotly

